move record summary to utils, update related changes
